+++ package.json
@@ -15,8 +15,8 @@
     "@types/jest": "^26.0.15",
     "@types/node": "^15.12.4",
     "@types/react": "^17.0.11",
-    "animejs": "^3.2.1",
     "@types/react-dom": "^17.0.8",
+    "animejs": "^3.2.1",
     "react": "^17.0.2",
     "react-copy-to-clipboard": "^5.0.3",
     "react-dom": "^17.0.2",
@@ -25,10 +25,10 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-transition-group": "^4.4.2",
-    "typescript": "^4.3.4",
-    "styled-components": "^5.3.0",
     "simple-peer": "^9.11.0",
     "socket.io-client": "^4.1.2",
+    "styled-components": "^5.3.0",
+    "typescript": "^4.3.4",
     "web-vitals": "^2.0.1"
   },
   "scripts": {
@@ -61,9 +61,10 @@
   "devDependencies": {
     "@types/animejs": "^3.1.3",
     "@types/loadable__component": "^5.13.3",
+    "@types/react-copy-to-clipboard": "^5.0.0",
     "@types/react-router-dom": "^5.1.7",
-    "@types/styled-components": "^5.1.9",
     "@types/simple-peer": "^9.11.0",
+    "@types/styled-components": "^5.1.9",
     "gh-pages": "^3.2.0"
   },
   "license": "MIT"
