+++ src/components/gridview/VideoFeed.tsx
@@ -1,19 +1,18 @@
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
 import { useContext } from "react";
 import { SocketContext } from "../context/Context";
 import VideoPlayer from "../videoplayer/VideoPlayer";
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    video_feed: {
-      flexGrow: 1,
-      width: "49%",
-      background: "#2F4F4F",
-      border: "6px solid #000000",
-      box_sizing: "border-box",
-    },
-  })
-);
+// const useStyles = makeStyles((theme: Theme) =>
+//   createStyles({
+//     video_feed: {
+//       flexGrow: 1,
+//       width: "49%",
+//       background: "#2F4F4F",
+//       border: "6px solid #000000",
+//       box_sizing: "border-box",
+//     },
+//   })
+// );
 
 const VideoFeed = () => {
   const context = useContext(SocketContext);
@@ -24,7 +23,7 @@ const VideoFeed = () => {
     //  , call
   } = context;
 
-  const classes = useStyles();
+  // const classes = useStyles();
   return <>{stream && <VideoPlayer stream={stream!} />}</>;
 };
